trainer/requirements.txt
# Base Python packages for the 3dgs_app training environment.
# Install order:
# 1) pip install -r trainer/requirements-cu118.txt
# 2) pip install -r trainer/requirements.txt
# 3) pip install --force-reinstall --no-deps --index-url https://docs.gsplat.studio/whl/pt21cu118 -r trainer/requirements-gsplat-cu118.txt
nerfstudio==1.1.5
numpy==1.26.4
open3d==0.19.0
pillow==12.2.0
plyfile==1.1.3
scipy==1.15.3
tensorboard==2.20.0
tyro==1.0.12
viser==0.2.7
